+++ common/Utils.lua
@@ -1,12 +1,32 @@
 local addonName, AUR = ...
 
-local L = AUR.localization
+local L = AUR.Localization
+
+local AWL = ArcaneWizardLibrary
 
 local Utils = {}
 
-----------------------
---- Local Funtions ---
-----------------------
+-----------------------
+--- Local Functions ---
+-----------------------
+
+local function CopyTable(source)
+	local target = {}
+
+	for key, value in pairs(source) do
+		if type(value) == "table" then
+			target[key] = CopyTable(value)
+		else
+			target[key] = value
+		end
+	end
+
+	return target
+end
+
+local function GetCharacterRealmKey()
+	return AWL.Utils:GetCharacterRealmKey()
+end
 
 function Utils:GetToday()
 	return date("%Y-%m-%d")
@@ -23,12 +43,12 @@ function Utils:GetCharacterInfo()
 	return realm, char
 end
 
----------------------
---- Main Funtions ---
----------------------
+------------------------
+--- Public Functions ---
+------------------------
 
 function Utils:PrintDebug(msg)
-	if AUR.options.other["debug-mode"] then
+	if AUR.settings.general["debug-mode"] then
 		DEFAULT_CHAT_FRAME:AddMessage(ORANGE_FONT_COLOR:WrapTextInColorCode(addonName .. " (Debug): ")  .. msg)
 	end
 end
@@ -37,49 +57,103 @@ function Utils:PrintMessage(msg)
 	DEFAULT_CHAT_FRAME:AddMessage(NORMAL_FONT_COLOR:WrapTextInColorCode(addonName .. ": ") .. msg)
 end
 
+function Utils:IsAccountProfile()
+	local characterRealmKey = GetCharacterRealmKey()
+
+	return Aurarium_Options_v3.profileKeys[characterRealmKey]["use-account"]
+end
+
+function Utils:OpenSettingsOnLoading()
+	local characterRealmKey = GetCharacterRealmKey()
+
+	if Aurarium_Options_v3.profileKeys[characterRealmKey]["open-settings"] then
+		Settings.OpenToCategory(AUR.MAIN_CATEGORY_ID)
+
+		Aurarium_Options_v3.profileKeys[characterRealmKey]["open-settings"] = false
+	end
+end
+
+function Utils:ToggleProfileMode()
+	local characterRealmKey = GetCharacterRealmKey()
+	local useAccountProfile = self:IsAccountProfile()
+
+	Aurarium_Options_v3.profileKeys[characterRealmKey]["use-account"] = not useAccountProfile
+	Aurarium_Options_v3.profileKeys[characterRealmKey]["open-settings"] = true
+end
+
+function Utils:ResetAllCharacterProfiles()
+	local characterRealmKey = GetCharacterRealmKey()
+
+	Aurarium_Options_v3.profiles = {}
+	Aurarium_Options_v3.profileKeys = {}
+
+	Aurarium_Options_v3.profileKeys[characterRealmKey] = {
+		["use-account"] = true,
+		["open-settings"] = true
+	}
+end
+
 function Utils:InitializeDatabase()
 	local realm, char = Utils:GetCharacterInfo()
+	local characterRealmKey = GetCharacterRealmKey()
 
-	if (not Aurarium_Options_v2) then
-		Aurarium_Options_v2 = {
-			["general"] = {
-				["minimap-button"] = {
-					["hide"] = false
-				}
-			},
-			["currency-overview"] = {},
-			["other"] = {}
+	local defaults = {
+		["general"] = {
+			["minimap-button"] = {
+				["hide"] = false
+			}
+		},
+		["currency-overview"] = {}
+	}
+
+	if not Aurarium_Options_v3 then
+		Aurarium_Options_v3 = {
+			["account"] = CopyTable(defaults),
+			["profiles"] = {},
+			["profileKeys"] = {}
 		}
 	end
 
-	if (not Aurarium_DataDates) then
+	if not Aurarium_Options_v3.profiles[characterRealmKey] then
+		Aurarium_Options_v3.profiles[characterRealmKey] = CopyTable(defaults)
+	end
+
+	if not Aurarium_Options_v3.profileKeys[characterRealmKey] then
+		Aurarium_Options_v3.profileKeys[characterRealmKey] = {
+			["use-account"] = true,
+			["open-settings"] = false
+		}
+	end
+
+	if Aurarium_Options_v3.profileKeys[characterRealmKey]["use-account"] then
+		AUR.settings.general = Aurarium_Options_v3.account["general"]
+		AUR.settings.currencyOverview = Aurarium_Options_v3.account["currency-overview"]
+	else
+		AUR.settings.general = Aurarium_Options_v3.profiles[characterRealmKey]["general"]
+		AUR.settings.currencyOverview = Aurarium_Options_v3.profiles[characterRealmKey]["currency-overview"]
+	end
+
+	if not Aurarium_DataDates then
 		Aurarium_DataDates = {}
 	end
 
-	if (not Aurarium_DataCharacter) then
+	if not Aurarium_DataCharacter then
 		Aurarium_DataCharacter = {}
 	end
 
-	if (not Aurarium_DataBalance) then
+	if not Aurarium_DataBalance then
 		Aurarium_DataBalance = {}
 	end
 
-	AUR.options = {}
-	AUR.options.general = Aurarium_Options_v2["general"]
-	AUR.options.currencyOverview = Aurarium_Options_v2["currency-overview"]
-	AUR.options.other = Aurarium_Options_v2["other"]
-
-	AUR.data = {}
 	AUR.data.dates = Aurarium_DataDates
 	AUR.data.character = Aurarium_DataCharacter
 	AUR.data.balance = Aurarium_DataBalance
 
-	AUR.data.character[realm] =  AUR.data.character[realm] or {}
-	AUR.data.character[realm][char] =  AUR.data.character[realm][char] or {}
+	AUR.data.character[realm] = AUR.data.character[realm] or {}
+	AUR.data.character[realm][char] = AUR.data.character[realm][char] or {}
 
-	AUR.data.balance =  AUR.data.balance or {}
-	AUR.data.balance[realm] =  AUR.data.balance[realm] or {}
-	AUR.data.balance[realm][char] =  AUR.data.balance[realm][char] or {}
+	AUR.data.balance[realm] = AUR.data.balance[realm] or {}
+	AUR.data.balance[realm][char] = AUR.data.balance[realm][char] or {}
 
 	if AUR.GAME_TYPE_MAINLINE then
 		AUR.data.balance["Warband"] = AUR.data.balance["Warband"] or {}
@@ -93,10 +167,10 @@ function Utils:InitializeMinimapButton()
 		icon     = AUR.MEDIA_PATH .. "icon-round.blp",
 		OnClick  = function(self, button)
 			if button == "LeftButton" then
-				if AUR.overview:IsShown() then
-					AUR.overview:Hide()
+				if AUR.modules.Overview:IsShown() then
+					AUR.modules.Overview:Hide()
 				else
-					AUR.overview:Show()
+					AUR.modules.Overview:Show()
 				end
 			elseif button == "RightButton" then
 				if not InCombatLockdown() then
@@ -115,7 +189,7 @@ function Utils:InitializeMinimapButton()
 	})
 
 	self.minimapButton = LibStub("LibDBIcon-1.0")
-	self.minimapButton:Register("Aurarium", LDB, AUR.options.general["minimap-button"])
+	self.minimapButton:Register("Aurarium", LDB, AUR.settings.general["minimap-button"])
 end
 
-AUR.utils = Utils
+AUR.modules.Utils = Utils