+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Aurarium/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Aurarium/releases)
 
+**Important note: All settings have been reset to default values.**
+
+- Added: Profile settings with account profile and character profile selection
+- Updated: deDE, enUS, ruRU localization
 - Minor code adjustments